batusdt_bot/requirements.txt
python-telegram-bot==12.6.1
pandas==1.1.5
python-binance==0.7.5